ci: add CI workflows for publishing and promotion
This commit adds 4 Github workflows:

`push.yaml` runs when there are pushes to the
`main` branch, builds the snap and uploads it to
the Snap store automatically, releasing the built
artifact to the `latest/candidate` channel.

`pull-request.yaml` executes on Pull Requests, and
simply checks the Snap builds correctly.

`update-snap.yaml` runs at 10:00 UTC every day,
watching for new releases, and automatically
committing and updated `snapcraft.yaml` to
the `main` branch, assuming that the Snap builds.

`promote.yaml` is a manually dispatched workflow
that provides a convenience mechanism for promoting
from `candidate` -> `stable`.

Note that the Github Actions settings will need to
be altered such that the Github Token can write
changes to the repo, and there will need to be a
Secret created named `STORE_TOKEN` which can be
generated using `snapcraft export-login`, and needs
permissions to manage the `zellij` snap on both
the `latest/candidate` and `latest/stable` tracks.